feat: Recurse into target paths if -r is given
Adding -r lets the user specify a top level folder which we then descend
into recursively.

Adding a large folder structure would have made the program grind to a
halt before showing the rendered GUI. To avoid this, a safety limit of
1000 total files is introduced (for both recursive and non-recursive
finding of images). Users will probably not notice this,
since handling that many files manually would be a tedious task.

By ignoring errors, we won't descend into those directories, but we will
descend into everything else. This let's us cast a wide net with the
input directories, and handle it optimistically.